01-search-tool-agent/requirements.txt
azure-ai-projects>=2.0.0
azure-identity>=1.21.0
azure-search-documents>=11.6.0
requests>=2.32.0
